Clear native queries. (#811)
* Clear native queries.

* Fix code smells

* Fix code smells

* Fix code smells

+++ service/src/main/java/greencity/service/ubs/OrdersAdminsPageServiceImpl.java
@@ -224,9 +224,9 @@ public class OrdersAdminsPageServiceImpl implements OrdersAdminsPageService {
         List<OptionForColumnDTO> optionForColumnDTOS = new ArrayList<>();
         OrderStatus[] orderStatuses = OrderStatus.values();
         for (OrderStatus o : orderStatuses) {
-            String ua = orderStatusTranslationRepository.getOrderStatusTranslationById(o.getNumValue())
+            String ua = orderStatusTranslationRepository.getOrderStatusTranslationById((long) o.getNumValue())
                 .orElseThrow(() -> new EntityNotFoundException("Order status have not found")).getName();
-            String en = orderStatusTranslationRepository.getOrderStatusTranslationById(o.getNumValue())
+            String en = orderStatusTranslationRepository.getOrderStatusTranslationById((long) o.getNumValue())
                 .orElseThrow(() -> new EntityNotFoundException("Order status have not found")).getName();
             optionForColumnDTOS
                 .add(OptionForColumnDTO.builder().key(o.toString()).ua(ua).en(en).filtered(false).build());
